chapter3-1-1/src/test/java/com/springboot/chapter311/webApi/UserControllerTest.java
package com.springboot.chapter311.webApi;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.internal.matchers.Equals;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.RequestBuilder;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import static org.hamcrest.Matchers.equalTo;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringRunner.class)
@SpringBootTest
public class UserControllerTest {
    private MockMvc mvc;

    @Before
    public void setUp() {
        mvc = MockMvcBuilders.standaloneSetup(
                new HelloController(), new UserController()
        ).build();
    }

    @Test
    public void getIndex() throws Exception {
        mvc.perform(MockMvcRequestBuilders.get("/index").accept(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(content().string(equalTo("index")));
    }

    @Test
    public void testUsers() throws Exception {
        RequestBuilder request = null;
        /** 1.获取用户信息(结果为空)**/
        request = get("/users/getUserList");
        mvc.perform(request)
                .andExpect(status().isOk())
                .andExpect(content().string(equalTo("[]")));
        /** 2.写入用户信息 **/
        request = post("/users/setUser")
                .param("id", "1")
                .param("name", "冰棍")
                .param("age", "18");
        mvc.perform(request)
                .andExpect(status().isOk())
                .andExpect(content().string(equalTo("success")));
        /** 3.获取用户信息(结果不为空)**/
        request = get("/users/getUserList");
        mvc.perform(request)
                .andExpect(status().isOk())
                .andExpect(content().string(equalTo("[{\"id\":1,\"name\":\"冰棍\",\"age\":18}]")));
        /** 4.put修改id为1的user**/
        request = put("/users/updateUserInfo/1")
                .param("name", "sober")
                .param("age", "30");
        mvc.perform(request)
                .andExpect(status().isOk())
                .andExpect(content().string(equalTo("success")));
        /** 5.获取ID为1的用户信息**/
        request = get("/users/getUserInfoById/1");
        mvc.perform(request)
                .andExpect(status().isOk())
                .andExpect(content().string(equalTo("{\"id\":1,\"name\":\"sober\",\"age\":30}")));
        /** 6.删除用户信息 **/
        request = delete("/users/delete/1");
        mvc.perform(request)
                .andExpect(status().isOk())
                .andExpect(content().string(equalTo("success")));
        /** 7.getUserList**/
        request = get("/users/getUserList");
        mvc.perform(request)
                .andExpect(status().isOk())
                .andExpect(content().string(equalTo("[]")));
    }
}
